Moving lambda util functions
