+++ lambdas/retry/retry_lambda.py
@@ -1,0 +1,56 @@
+"""
+Lambda function to attempt to retry executing failed ETL processes,
+giving configurable retry logic such as 
+maximum retry attempts (default: 5) or backoff timing.
+"""
+
+import json
+import logging
+import os
+
+from lambdas.lambda_triggers_etl import lambda_triggers_etl
+
+logger = logging.getLogger()
+logger.setLevel("INFO")
+
+MAXIMUM_RETRY_ATTEMPTS = 5
+BACKOFF_TIMING_SCHEDULE = None
+
+def lambda_handler(event: dict, context):
+    """
+    Lambda function to detect when an ETL process has failed, 
+    then attempt to retry the process while iterating on 
+    retry variables (retry attempts and exponential backoff).
+    """
+    try:
+
+        logger.info("Retry lambda checking ETL process status.")
+
+        event_status = get_event_status_type(event)
+        if event_status == "FAILED":
+            # invoke etl lambda again
+            logger.info(f"ETL process failure detected, retrying")
+
+
+            
+        else:
+            # ETL process did not fail, do nothing
+            logger.info("ETL process status normal, continuing")
+            pass
+    except Exception as e:
+        logger.error(f"Error occured while attempting to retry ETL process. Error: {e}")
+        raise e
+
+
+def get_event_status_type(event):
+    try:
+        if isinstance(event, str):
+            event = json.loads(event)
+        
+        if "event_type" in event:
+            return event["event_type"]
+
+        
+    except Exception as e:
+        logger.error(f"Error occured while retrieving ETL process status: {e}")
+        raise e
